Move engine utils to utils.
